requirements.txt
jupyter==1.0.0
jupyterlab==2.2.2
matplotlib==3.2.1
numpy==1.18.3
pandas==1.0.3
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.1
PyYAML==5.3.1